+++ homeassistant/components/google_assistant/data_redaction.py
@@ -5,7 +5,7 @@ from collections.abc import Callable
 from typing import Any
 
 from homeassistant.core import callback
-from homeassistant.util.redact import async_redact_data, partial_redact
+from homeassistant.helpers.redact import async_redact_data, partial_redact
 
 SYNC_MSG_TO_REDACT: dict[str, Callable[[str], str]] = {
     "agentUserId": partial_redact,
